Moved Tutorial to screens folder and removed style declaration for Cool Bottom from the function
