+++ components/CoolButton.tsx
@@ -7,20 +7,21 @@ export type CoolButton = {
 	style?: TouchableOpacityProps | ViewProps | FlexStyle,
 }
 export default function CoolButton(props: CoolButton) {
-	const styles = StyleSheet.create({
-		buttons: {
-			borderStyle: "dotted",
-			borderWidth: 1,
-			padding: 20,
-		},
-		text: {
-			fontSize: 30,
-			textAlign: 'center',
-			fontFamily: "Nunito"
-		}
-	})
 	return (
 		<TouchableOpacity style={[styles.buttons, props.style]} onPress={props.onPress}>
 			<CoolText style={styles.text}>{props.text}</CoolText>
 		</TouchableOpacity>)
 }
+
+const styles = StyleSheet.create({
+	buttons: {
+		borderStyle: "dotted",
+		borderWidth: 1,
+		padding: 20,
+	},
+	text: {
+		fontSize: 30,
+		textAlign: 'center',
+		fontFamily: "Nunito"
+	}
+})
